# Speed up group mutate operations