@mikelove 2025-03-18
Thanks @ppaxisa, a big speedup!

I'm thinking of modifying the message from

```
metadata contains S4 columns, using plyranges group operations.
You can speed up group operations by coercing metadata columns to S3
```

to something like:

```
metadata contains S4 columns, using lapply group operations.
removing S4 columns will speed up group operations.
```

Thoughts?

I'm not sure S3 is accurate here.