@ppaxisa 2025-03-18
Yes S3 is not really accurate actually, I did not think too much about the message content tbh. The edited version looks good to me.